Make runserver command more flexible
Since I quite often catch myself cd-ing to the correct directory to
start the server, because I am in a subdir of a Django project; change
to the correct directory automatically before starting the server.

It may sometimes be needed to make the server listen at 0.0.0.0 so
optionally override the address with a second argument to runserver.