Install baps1 (to ~/bin)
baps1 is used in the prompt and shows the current tty and time since the
last call to baps1 for that tty, which can give you a rough estimation
of the time duration of the last command. It is basically like calling
every command with `time <command>` (plus the time it took you to write
out the command).

Example output: `p1:5m43s`

baps1 is written by @ryuslash (who gets mailbombed everytime I mention
him :grinning:).