Add git-remote-gh script
usage: git-remote-gh <user/repo> [<remote>] [<branch>]
       git-remote-gh [options]

Options:
-h, --help       show this help information
-v               show configured remotes

Configure remote urls for a public Github repository
(using https for fetch / ssh for push)

If <remote> is not specified, it will default to 'origin.'
If <branch> is not specified, it will default to the name of
the current local branch (git branch --show-current).

This configuration makes it possible to fetch and pull changes
without the need for an ssh-agent with an ssh-key loaded and/or
having to type a password. It also sets the upstream remote
tracking branch without pushing any data.

Fetch urls use the https scheme "https://github.com/<user/repo>".
Pull urls use the ssh scheme "git@github.com:<user/repo>".
The repository should be public and you should have push rights.

Scripts in PATH that start with git- will be made available
as git subcommands. It is therefore possible to run this script
as 'git remote-gh'. Be aware however that invoking
'git remote-gh --help' will try to find a manpage instead
and fail instead of showing this help.