Add new `tmux-swprefix` shell script to bin
usage: tmux-swprefix [option]

Helper for using tmux over ssh in tmux (over ssh in tmux).

Switch the prefix key of a running tmux session and indicate
it by updating the background color of the status bar.

Options:
-h, --help      show this help information
-V, --version   show version information
-1, -b,         switch to preset 1 / set prefix key to C-b
-2, -a,         switch to preset 2 / set prefix key to C-a
-3, -t,         switch to preset 3 / set prefix key to C-t
-n, --next      cycle through presets / prefix keys

\# Why use this?
When you are in a tmux session and want to ssh into a server
and use tmux there, it is easily enough to change the
prefix-key in '.tmux.conf' on one of the machines so they
don't clash with eachother by responding to the same key
commands.

But when you add another layer of tmux/ssh/server or like to
mix things up, this can get tricky. The config is rigid in the
way that it can support *only one* of multiple ways that these
machines are able to be linked up with eachother.

This script is created to have a flexible way to "nest" tmux
though ssh, 3 layers deep. It sets the prefix key of a running
tmux session and indicates it by updating the background color
of the status bar.